+++ vector/src/main/java/im/vector/app/features/settings/devtools/AccountDataCreateController.kt
@@ -8,6 +8,8 @@
 package im.vector.app.features.settings.devtools
 
 import com.airbnb.epoxy.TypedEpoxyController
+import im.vector.app.R
+import im.vector.app.core.epoxy.checkBoxItem
 import im.vector.app.core.resources.StringProvider
 import im.vector.app.core.ui.list.genericFooterItem
 import im.vector.app.features.form.formEditTextItem
@@ -36,11 +38,20 @@ class AccountDataCreateController @Inject constructor(
         }
         formEditTextItem {
             id("account_data_type")
+            layout(R.layout.item_form_text_input_compact)
             enabled(true)
             value(data.draft.type)
             hint(host.stringProvider.getString(CommonStrings.dev_tools_form_hint_type))
             onTextChange { text ->
                 host.interactionListener?.processAction(AccountDataAction.DraftTypeChange(text))
+            }
+        }
+        checkBoxItem {
+            id("account_data_encrypt")
+            title(host.stringProvider.getString(CommonStrings.account_data_encrypt))
+            checked(data.draft.encrypt)
+            checkChangeListener { _, checked ->
+                host.interactionListener?.processAction(AccountDataAction.DraftEncryptChange(checked))
             }
         }
         formMultiLineEditTextItem {